Some Checks for Strassen's Algorithm
This file includes 4 new functions.
1. CheckPowerofTwo
2. MakePowerofTwo
3.arrayCheck
4.resizeArray
First 2 functions are used to check that whether the size of matrix is power of 2 or not. If not, the size is converted into power of 2. arrayCheck checks that whether size of rows and columns are same or not. If not, take larger value, check it for power of 2 and pass it to resizeArray. resizeArray will increase the size, existing data will set accordingly and remaining spaces will be filled with zeros. e.g. the matrix 
{1,2,3
4,5,6
7,8,9} 
resizes as
{1,2,3,4
5,6,7,8
9,0,0,0,
0,0,0,0}